% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/awiData.R
\name{getAwiSet}
\alias{getAwiSet}
\title{Return the \emph{National Average Wage Index} data set.}
\usage{
getAwiSet(endYear = NULL)
}
\arguments{
\item{endYear}{the last year to include in the returned data set}
}
\value{
the AWI data set
}
\description{
The \emph{National Average Wage Index} is used to normalize wages across
different years. Wage indexing depends on the year in which a person is
first elgible to receive OASDI benefits. The year of first elgibility for
OASDI is at age 62. An individuals earnings are always indexed to the average
wage level two years prior to the year of first eligibility. This is called
the indexing year. Thus, for a person reaching age 62 in 2017, the person’s
earnings would be indexed to the average wage index for 2015, which is the
indexing year. A table of \href{https://www.ssa.gov/OACT/COLA/awiseries.html}{Average Wage Indexes}
is available on the Social Security web site. This data set comes from there,
but it is not a real time fetch, so the data is current up to the time that
this package was created. In order to determine the Average Indexed Monthly
Earnings when the indexing year is beyond the last year in the AWI data, the
data is fit with a model which is used to project future wage index values.
If no endYear is provided the function returns the data set of measured values
without any projected values.
}
\examples{
getAwiSet(endYear = 2025)
getAwiSet()

}
