% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/awiData.R
\name{getAwiValues}
\alias{getAwiValues}
\title{Return the \emph{National Average Wage Index} values for input years.}
\usage{
getAwiValues(years = NULL)
}
\arguments{
\item{year}{the years fof the AWI values to return}
}
\value{
the AWI values for \code{years}
}
\description{
Return the AWI values for input years. The values are either
measured or predicted through a model depending on whether or
not a measure value exists for the input year. Predicted values
will only be produced for years later than the last year in the
data set downloaded from the government site. No values for years
before the first year in the set will be predicted.
}
\examples{
getAwiValues(years = 2025)
getAwiValues(years = 2000:2025)

}
