% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benefitBasesData.R
\name{getBenefitBasesSet}
\alias{getBenefitBasesSet}
\title{Return the \emph{Benefit Bases} data set.}
\usage{
getBenefitBasesSet(endYear = NULL)
}
\arguments{
\item{endYear}{the last year to include in the returned data set}
}
\value{
the Benefit Bases data set
}
\description{
Each Social Security benefit is based on the history of earned and
taxable income for the beneficiary (or the beneficiaries spouse).
The income considered is limited by the Benefit and Contribution Base
which changes from year to year to reflect the average of wages
earned in that year. This amount is called the Taxed Social Security
Earnings and it is related to the Average Wage Index. The Social
Security web site contains a table of \href{https://www.ssa.gov/oact/COLA/cbb.html}{Contribution and Benefits Bases}.
This data set comes from there, but it is not a real time fetch, so
the data is current up to the time that this package was created. In
order to determine the Benefit Base for a year that is beyond the
last year in the Benefit Bases data, the data is fit with a model
which is used to project future Benefit Base values. If no endYear
is provided the function returns the data set of measured values
without any projected values.
}
\examples{
getBenefitBasesSet(endYear = 2025)
getBenefitBasesSet()

}
