% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benefitBasesData.R
\name{getBenefitBaseValues}
\alias{getBenefitBaseValues}
\title{Return the \emph{Benefit Bases} values for input years.}
\usage{
getBenefitBaseValues(years = NULL)
}
\arguments{
\item{year}{the years of the Benefit Base values to return}
}
\value{
the Benefit Base values for \code{years}
}
\description{
Return the Benefit Bases for input years. The values are either
measured or predicted through a model depending on whether or
not a measure value exists for the input year. Predicted values
will only be produced for years later than the last year in the
data set downloaded from the government site. No values for years
before the first year in the set will be predicted.
}
\examples{
getBenefitBaseValues(years = 2025)
getBenefitBaseValues(years = 2000:2025)

}
