% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{getColaSet}
\alias{getColaSet}
\title{Return the \emph{Cost of Living Adjustment} data set.}
\usage{
getColaSet(endYear = NULL)
}
\arguments{
\item{endYear}{the last year to include in the returned data set}
}
\value{
the cost of living adjustment data set
}
\description{
Legislation enacted in 1973 provides for cost-of-living adjustments,
or COLAs. With COLAs, Social Security and Supplemental Security Income
(SSI) benefits keep pace with inflation. The Social Security Act specifies
a formula for determining each COLA. According to the formula, COLAs are
based on increases in the Consumer Price Index for Urban Wage Earners and
Clerical Workers (CPI-W). CPI-Ws are calculated on a monthly basis by the
Bureau of Labor Statistics. A COLA effective for December of the current
year is equal to the percentage increase (if any) in the average CPI-W for
the third quarter of the current year over the average for the third quarter
of the last year in which a COLA became effective. If there is an increase,
it must be rounded to the nearest tenth of one percent. If there is no
increase, or if the rounded increase is zero, there is no COLA. For an
example computation of COLA, see the Lastest
\href{https://www.ssa.gov/oact/COLA/latestCOLA.html}{Cost of Living Adjustment}
page on the Social Security web site.
}
\examples{
getColaSet(endYear = 2025)
getColaSet()

}
