% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{getFamilyMax}
\alias{getFamilyMax}
\title{The Family Maximum Benefit is the maximum Social Security benefit
that can be paid based on an individual's earnings record. Like the
PIA, it is based on the average of the highest 35 years of indexed
earnings.}
\usage{
getFamilyMax(maxPiaBends = NULL, aime = NULL)
}
\arguments{
\item{maxPiaBends}{the family maximum PIA bend points from \code{getMaxPiaBendPts()}}

\item{aime}{The Average Indexed Monthly Value from \code{getAimeValue()}}
}
\value{
The Family Maximum Benefit
}
\description{
The Family Maximum Benefit is the maximum Social Security benefit
that can be paid based on an individual's earnings record. Like the
PIA, it is based on the average of the highest 35 years of indexed
earnings.
}
\examples{
getPia(piaBends = JQPublicBends, aime = JQPublicAime)

}
