% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statement.R
\name{getFicaEarnings}
\alias{getFicaEarnings}
\title{Return \emph{FICA} earnings data.}
\usage{
getFicaEarnings(statement = NULL, endYear = NULL)
}
\arguments{
\item{statement}{the Social Security Statement Object returned from \code{readStatement}}

\item{endYear}{the last year to include in the returned data set}
}
\value{
the FICA earnings data set
}
\description{
Taxes under the Federal Insurance Contributions Act (FICA) are
composed of the old-age, survivors, and disability insurance taxes,
also known as social security taxes, and the hospital insurance tax,
also known as Medicare taxes. Different rates apply for these taxes.
The amount an employee pays in FICA tax is capped at the Benefit and
Contribution Base which can change from year to year according to
the Average Wage Index (AWI). The data set returned will contain the
actual values for years that the government listed at the time
this package was built. For any year beyond that, an estimate is
returned based on estimated values for the Benefit and Contribution
Base and full wage earnings. Since Medicare taxes are not capped
the full wage earnings are estimated using the medicare earnings
(they are the same). The minimum of estimated Medicare earnings
and estimated Benefit and Contribution Base will be the FICA earnings
returned.
}
\details{
The \code{endYear} is the last year of FICA earnings to return. If no
\code{endYear} is given, only the FICA earnings contained in the input
Social Security Statement object are returned.
}
\examples{
getFicaEarnings(statement = myStatement, endYear = 2025)
getFicaEarnings(statement = myStatement)
}
