% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fraData.R
\name{getFraSet}
\alias{getFraSet}
\title{Return the \emph{Full Retirement Age} data set}
\usage{
getFraSet(endYear = NULL)
}
\arguments{
\item{endYear}{the last year to include in the returned data set}
}
\value{
the full retirement age data set
}
\description{
A worker's Full Retirement Age is the age at which the worker
is entitled to the full Primary Insurance Amount (PIA). If a
worker starts taking distributions before the full retirement
age, the distributions are reduced. Conversely, if a worker
delays distributions after the full retirement age, the
distributions are more.
}
\examples{
getFraSet(endYear = 2025)
getFraSet()

}
