% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fraData.R
\name{getFraValues}
\alias{getFraValues}
\title{Return the \emph{Full Retirement Age} values for input years.}
\usage{
getFraValues(years = NULL)
}
\arguments{
\item{year}{the years of the Full Retirement Age to return}
}
\value{
the Full Retirement Age for \code{years}
}
\description{
Return the Full Retirement Age for input years. No values for years
before the first year in the set will be predicted.
}
\examples{
getFraValues(years = 2025)
getFraValues(years = 2000:2025)

}
