% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{getMaxPiaBendPts}
\alias{getMaxPiaBendPts}
\title{Return the maximum family benefit \emph{Primary Insurance Amount Bend Points} for a year.}
\usage{
getMaxPiaBendPts(year = NULL)
}
\arguments{
\item{year}{the year for the family maximum PIA bend points to return}
}
\value{
the family maximum PIA bend points for \code{year}
}
\description{
Return the family maximum PIA bend points for a year. These values are
calculated using AWI values.
}
\examples{
getMaxPiaBendPts(year = 2025)

}
