% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statement.R
\name{getMedicareEarnings}
\alias{getMedicareEarnings}
\title{Return the \emph{Medicare} earnings data set.}
\usage{
getMedicareEarnings(statement = NULL, endYear = NULL, predict = FALSE)
}
\arguments{
\item{statement}{the Social Security Statement Object returned from \code{readStatement}}

\item{endYear}{the last year to include in the returned data set}

\item{predict}{if TRUE, use a MARS model to predict future earnings. Otherwise replicate
the last know earnings value (default)}
}
\value{
the Medicare earnings data set
}
\description{
Taxes under the Federal Insurance Contributions Act (FICA) are
composed of the old-age, survivors, and disability insurance taxes,
also known as social security taxes, and the hospital insurance tax,
also known as Medicare taxes. Different rates apply for these taxes.
The amount an employee pays in FICA tax is capped at the Benefit and
Contribution Base which can change from year to year according to
the Average Wage Index (AWI). The data set returned will contain the
actual values for years that the government listed at the time
this package was built. For any year beyond that, an estimate is
returned.
}
\details{
The \code{endYear} is the last year of Medicare earnings to return. If no
\code{endYear} is given, only the Medicare earnings contained in the input
Social Security Statement object are returned.
}
\examples{
getMedicareEarnings(statement = myStatement, endYear = 2025)
getMedicareEarnings(statement = myStatement)
}
