% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{getPia}
\alias{getPia}
\title{The Primary Insurance Amount (PIA) is the actual Social Security
benefit. It is based on the average of the highest 35 years of
indexed earnings.}
\usage{
getPia(piaBends = NULL, aime = NULL)
}
\arguments{
\item{piaBends}{the PIA bend points from \code{getPiaBendPts()}}

\item{aime}{The Average Indexed Monthly Value from \code{getAimeValue()}}
}
\value{
The Primary Insurance Amount
}
\description{
The Primary Insurance Amount (PIA) is the actual Social Security
benefit. It is based on the average of the highest 35 years of
indexed earnings.
}
\examples{
getPia(piaBends = JQPublicBends, aime = JQPublicAime)

}
