% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{getPiaAdjustments}
\alias{getPiaAdjustments}
\title{Return the \emph{Primary Insurance Amount Adjustments} for a birth year.}
\usage{
getPiaAdjustments(birthYear = NULL)
}
\arguments{
\item{birthYear}{the birth year of the worker}
}
\value{
the percentage adjustment data set for the input birth year
}
\description{
If a worker takes benefit distributions before or after the
normal retirement age, the PIA is adjusted by multiplying it
by a percentage.
A \href{https://www.ssa.gov/oact/ProgData/ar_drc.html}{table of adjustments}
is available on the Social Security web site. This function returns
the adjustments listed in the table for the input birth year.
}
\examples{
getPiaAdjustmentSet(birthYear = 1955)

}
