% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{getPiaAdjustmentsSet}
\alias{getPiaAdjustmentsSet}
\title{Return the \emph{Primary Insurance Amount Adjustments} data set.}
\usage{
getPiaAdjustmentsSet(endYear = NULL)
}
\arguments{
\item{birthYear}{the last birth year to include}
}
\value{
the percentage adjustment data set for all ages up to the birth year
}
\description{
If a worker takes benefit distributions before or after the
normal retirement age, the PIA is adjusted by multiplying it
by a percentage.
A \href{https://www.ssa.gov/oact/ProgData/ar_drc.html}{table of adjustments}
is available on the Social Security web site. Any year of birth greater
than 1960 uses the same percentage adjustments as for 1960. Therefore
any year passed into the function greater than 1960 will return the
same data set.
}
\examples{
getPiaAdjustmentSet(endYear = 1955)
getPiaAdjustmentSet()

}
