% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{getPiaBendPts}
\alias{getPiaBendPts}
\title{Return the \emph{Primary Insurance Amount Bend Points} for a year.}
\usage{
getPiaBendPts(year = NULL)
}
\arguments{
\item{year}{the year for the PIA bend points to return}
}
\value{
the PIA bend points for \code{year}
}
\description{
Return the PIA bend points for a year. These values are
calculated using AWI values.
}
\examples{
getPiaBendPts(year = 2025)

}
