% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{getPiaBendPtsSet}
\alias{getPiaBendPtsSet}
\title{Return the \emph{Primary Insurance Amount Bend Points} data set.}
\usage{
getPiaBendPtsSet(endYear = NULL)
}
\arguments{
\item{endYear}{the last year to include in the returned data set}
}
\value{
the primary insurance amount bend point data set
}
\description{
The Primary Insurance Amount (PIA) is the base value of a worker's OASDI
at full retirement age. It is calculated by summing three separate percentages
of portions of the worker's Average Indexed Monthly earnings (AIME). The
portions depend on the year in which a worker attains age 62, becomes
disabled before age 62, or dies before attaining age 62. The threshold
points that determine the portions of a worker's AIME that apply to each
percentage are called the PIA bend points. PIA bend points are based on
the Average Wage Index (AWI), using the AWI for 1977 as the reference year.
The Social Security web site shows \href{https://www.ssa.gov/oact/COLA/piaformula.html}{an example}
of how to calculate these points.
}
\examples{
getPiaBendPtsSet(endYear = 2025)
getPiaBendPtsSet()

}
