% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retirement.R
\name{getPIAData}
\alias{getPIAData}
\title{Return estimates individual's retirement benefits.}
\usage{
getPIAData(retData = NULL, ssStmt = NULL, retireAge = NULL)
}
\arguments{
\item{ssStmt}{the Social Security Statement as retrieved through read_xml_ssStatement()}

\item{retireAge}{the age of reciepient when earnings end. If no age is given, full retirement age (FRA) will be used.}

\item{retDate}{list of retirement related values as retrieved through getRetirementData}
}
\value{
Named list of several retirement related values including estimated benefit beginning at various ages
}
\description{
Provides an estimate of the monthly primary insurance amount (PIA) for benefits beginning at several ages
based on data from a Social Security Statement (normal retirement) are calculated based on AIME using a specific
retirement age (actually the age when benefits are started). Several retirement related variables
are calculated and returned in a named list. This data can be used to calculate estimated
primary insurance amount (PIA) if benefits begin at various ages (See getPIAData in this package).
}
