% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statement.R
\name{readStatement}
\alias{readStatement}
\title{Return Social Security information from an XML file.}
\usage{
readStatement(file = NULL)
}
\arguments{
\item{file}{the name of the XML file to read}
}
\value{
a list containing the Social Security Statement information
}
\description{
Your Social Security statement is available on the Social
Security web site as a PDF file and an XML file. This
function reads the XML file and returns a list containing
all the information the XML file provides. First download
the XML file you your computer then provide the file name
as an input to this function. Use either the full path name
or a relative path name as part of the file name input.
}
\examples{
read_xml_ssStatement(file = "./myStatementData/richieRich.xml")

}
